apps/chatserver/pyproject.toml
[project]
name = "warehouse-chatserver"
version = "0.1.0"
description = "Warehouse operations chat server backed by ClickHouse via MCP"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.20.0,<1",
    "clickhouse-connect>=0.8.17,<1",
    "fastapi>=0.117.1,<1",
    "fastmcp==2.12.3",
    "openai>=2.9.0,<3",
    "pydantic>=2.10.0,<3",
    "pydantic-ai==1.34.0",
    "pydantic-settings>=2.10.0,<3",
    "python-socketio>=5.11.4,<6",
    "python-ulid>=3.1.0,<4",
    "uvicorn[standard]>=0.37.0,<1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0,<9",
    "pytest-asyncio>=0.24.0,<1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["warehouse_chatserver"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
